Split ToolSet and EntityIconTool
