+++ srcjs/swiss-army-knife-card.js
@@ -34,12 +34,10 @@ import {
   LitElement, html, css, svg, unsafeCSS,
 } from 'lit-element';
 
-import { classMap } from 'lit-html/directives/class-map.js';
 import { styleMap } from 'lit-html/directives/style-map.js';
 import { unsafeSVG } from 'lit-html/directives/unsafe-svg.js';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import { selectUnit } from '@formatjs/intl-utils';
-import { fireEvent, stateIcon } from 'custom-card-helpers';
 import { version } from '../package.json';
 
 import {
@@ -57,26 +55,7 @@ import {
 import Merge from './merge';
 import Utils from './utils';
 import Templates from './templates';
-import BaseTool from './basetool';
-import CircularSliderTool from './circularslidertool';
-import RangeSliderTool from './rangeslidertool';
-import CircleTool from './circletool';
-import LineTool from './linetool';
-import SwitchTool from './switchtool';
-import RegPolyTool from './regpolytool';
-import UserSvgTool from './usersvg';
-import EllipseTool from './ellipsetool';
-import RectangleTool from './rectangletool';
-import RectangleToolEx from './rectangletoolex';
-import BadgeTool from './badgetool';
-// import EntityIconTool from './entityicontool'; // Uses SAK as cache. Can't split!
-import EntityStateTool from './entitystatetool';
-import EntityAreaTool from './entityareatool';
-import EntityNameTool from './entitynametool';
-import TextTool from './texttool';
-import SparklineBarChartTool from './sparklinebarcharttool';
-import SegmentedArcTool from './segmentedarctool';
-import HorseshoeTool from './horseshoetool';
+import Toolset from './toolset';
 
 // Original injector is buggy. Use a patched version, and store this local...
 import * as SvgInjector from '../dist/SVGInjector.min.js'; // lgtm[js/unused-local-variable]
@@ -86,690 +65,6 @@ console.info(
   'color: yellow; font-weight: bold; background: black',
   'color: white; font-weight: bold; background: dimgray',
 );
-
-// ++ Can't split these yet, due to circular dependencies...
-
-/** ****************************************************************************
-  * EntityIconTool class
-  *
-  * Summary.
-  *
-  */
-
-class EntityIconTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_ICON_CONFIG = {
-      classes: {
-        tool: {
-          'sak-icon': true,
-          hover: true,
-        },
-        icon: {
-          'sak-icon__icon': true,
-        },
-      },
-      styles: {
-        icon: {
-        },
-      },
-    };
-    super(argToolset, Merge.mergeDeep(DEFAULT_ICON_CONFIG, argConfig), argPos);
-
-    // from original
-    // this.config.entity = this.config.entity ? this.config.entity : 0;
-
-    // get icon size, and calculate the foreignObject position and size. This must match the icon size
-    // 1em = FONT_SIZE pixels, so we can calculate the icon size, and x/y positions of the foreignObject
-    // the viewport is 200x200, so we can calulate the offset.
-    //
-    // NOTE:
-    // Safari doesn't use the svg viewport for rendering of the foreignObject, but the real clientsize.
-    // So positioning an icon doesn't work correctly...
-
-    this.svg.iconSize = this.config.position.icon_size ? this.config.position.icon_size : 3;
-    this.svg.iconPixels = this.svg.iconSize * FONT_SIZE;
-
-    const align = this.config.position.align ? this.config.position.align : 'center';
-    const adjust = (align === 'center' ? 0.5 : (align === 'start' ? -1 : +1));
-
-    const clientWidth = 400; // testing
-    const correction = clientWidth / this._card.viewBox.width;
-
-    this.svg.xpx = this.svg.cx;
-    this.svg.ypx = this.svg.cy;
-
-    if (((this._card.isSafari) || (this._card.iOS)) && (!this._card.isSafari16)) {
-      this.svg.iconSize = this.svg.iconSize * correction;
-
-      this.svg.xpx = (this.svg.xpx * correction) - (this.svg.iconPixels * adjust * correction);
-      this.svg.ypx = (this.svg.ypx * correction) - (this.svg.iconPixels * 0.5 * correction) - (this.svg.iconPixels * 0.25 * correction);// - (iconPixels * 0.25 / 1.86);
-    } else {
-      // Get x,y in viewbox dimensions and center with half of size of icon.
-      // Adjust horizontal for aligning. Can be 1, 0.5 and -1
-      // Adjust vertical for half of height... and correct for 0.25em textfont to align.
-      this.svg.xpx = this.svg.xpx - (this.svg.iconPixels * adjust);
-      this.svg.ypx = this.svg.ypx - (this.svg.iconPixels * 0.5) - (this.svg.iconPixels * 0.25);
-    }
-    this.classes.icon = {};
-    this.styles.icon = {};
-    if (this.dev.debug) console.log('EntityIconTool constructor coords, dimensions, config', this.coords, this.dimensions, this.config);
-  }
-
-  /** *****************************************************************************
-  * EntityIconTool::_buildIcon()
-  *
-  * Summary.
-  * Builds the Icon specification name.
-  *
-  */
-  _buildIcon(entityState, entityConfig, toolIcon) {
-    return (
-      this.activeAnimation?.icon // Icon from animation
-      || toolIcon // Defined by tool
-      || entityConfig?.icon // Defined by configuration
-      || entityState?.attributes?.icon // Using entity icon
-      || stateIcon(entityState) // Use card helper logic (2021.11.21)
-    );
-  }
-
-  /** *****************************************************************************
-  * EntityIconTool::_renderIcon()
-  *
-  * Summary.
-  * Renders the icon using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the icon
-  *
-  * THIS IS THE ONE!!!!
-  */
-
-  _renderIcon() {
-    this.MergeAnimationClassIfChanged();
-    this.MergeAnimationStyleIfChanged();
-    this.MergeColorFromState(this.styles.icon);
-
-    const icon = this._buildIcon(
-      this._card.entities[this.defaultEntityIndex()],
-      this.config.hasOwnProperty('entity_index') ? this._card.config.entities[this.defaultEntityIndex()] : undefined,
-      this.config.icon,
-    );
-
-    if (true || (this.svg.xpx === 0)) {
-      this.svg.iconSize = this.config.position.icon_size ? this.config.position.icon_size : 2;
-      this.svg.iconPixels = this.svg.iconSize * FONT_SIZE;
-
-      // NEW NEW NEW Use % for size of icon...
-      this.svg.iconSize = this.config.position.icon_size ? this.config.position.icon_size : 2;
-      this.svg.iconPixels = Utils.calculateSvgDimension(this.svg.iconSize);
-
-      const align = this.config.position.align ? this.config.position.align : 'center';
-      const adjust = (align === 'center' ? 0.5 : (align === 'start' ? -1 : +1));
-
-      const clientWidth = 400;
-      const correction = clientWidth / (this._card.viewBox.width);
-
-      this.svg.xpx = this.svg.cx;// (x * this._card.viewBox.width);
-      this.svg.ypx = this.svg.cy;// (y * this._card.viewBox.height);
-
-      if (((this._card.isSafari) || (this._card.iOS)) && (!this._card.isSafari16)) {
-        // correction = 1; //
-        this.svg.iconSize = this.svg.iconSize * correction;
-        this.svg.iconPixels = this.svg.iconPixels * correction;
-
-        this.svg.xpx = (this.svg.xpx * correction) - (this.svg.iconPixels * adjust * correction);
-        this.svg.ypx = (this.svg.ypx * correction) - (this.svg.iconPixels * 0.9 * correction);
-        // - (this.svg.iconPixels * 0.25 * correction);// - (iconPixels * 0.25 / 1.86);
-        this.svg.xpx = (this.svg.cx * correction) - (this.svg.iconPixels * adjust * correction);
-        this.svg.ypx = (this.svg.cy * correction) - (this.svg.iconPixels * adjust * correction);
-      } else {
-        // Get x,y in viewbox dimensions and center with half of size of icon.
-        // Adjust horizontal for aligning. Can be 1, 0.5 and -1
-
-        this.svg.xpx = this.svg.cx - (this.svg.iconPixels * adjust);
-        this.svg.ypx = this.svg.cy - (this.svg.iconPixels * adjust);
-
-        if (this.dev.debug) console.log('EntityIconTool::_renderIcon - svg values =', this.toolId, this.svg, this.config.cx, this.config.cy, align, adjust);
-      }
-    }
-
-    if (!this.alternateColor) { this.alternateColor = 'rgba(0,0,0,0)'; }
-
-    if (!SwissArmyKnifeCard.sakIconCache[icon]) {
-      const theQuery = this._card.shadowRoot.getElementById('icon-'.concat(this.toolId))?.shadowRoot?.querySelectorAll('*');
-      if (theQuery) {
-        this.iconSvg = theQuery[0]?.path;
-      } else {
-        this.iconSvg = undefined;
-      }
-
-      if (!this.iconSvg) {
-      } else {
-        SwissArmyKnifeCard.sakIconCache[icon] = this.iconSvg;
-      }
-    } else {
-      this.iconSvg = SwissArmyKnifeCard.sakIconCache[icon];
-    }
-
-    let scale;
-
-    // NTS@20201.12.24
-    // Add (true) to force rendering the Safari like solution for icons.
-    // After the above fix, it seems to work for both Chrome and Safari browsers.
-    // That is nice. Now animations also work on Chrome...
-
-    if (this.iconSvg) {
-      // Use original size, not the corrected one!
-      this.svg.iconSize = this.config.position.icon_size ? this.config.position.icon_size : 2;
-      this.svg.iconPixels = Utils.calculateSvgDimension(this.svg.iconSize);
-
-      this.svg.x1 = this.svg.cx - this.svg.iconPixels / 2;
-      this.svg.y1 = this.svg.cy - this.svg.iconPixels / 2;
-      this.svg.x1 = this.svg.cx - (this.svg.iconPixels * 0.5);
-      this.svg.y1 = this.svg.cy - (this.svg.iconPixels * 0.5);
-
-      scale = this.svg.iconPixels / 24;
-      // scale = 1;
-      // Icon is default drawn at 0,0. As there is no separate viewbox, a transform is required
-      // to position the icon on its desired location.
-      // Icon is also drawn in a default 24x24 viewbox. So scale the icon to the required size using scale()
-      return svg`
-        <g id="icon-${this.toolId}" class="${classMap(this.classes.icon)}" style="${styleMap(this.styles.icon)}" x="${this.svg.x1}px" y="${this.svg.y1}px" transform-origin="${this.svg.cx} ${this.svg.cy}">
-          <rect x="${this.svg.x1}" y="${this.svg.y1}" height="${this.svg.iconPixels}px" width="${this.svg.iconPixels}px" stroke-width="0px" fill="rgba(0,0,0,0)"></rect>
-          <path d="${this.iconSvg}" transform="translate(${this.svg.x1},${this.svg.y1}) scale(${scale})"></path>
-        <g>
-      `;
-    } else {
-      // Note @2022.06.26
-      // overflow="hidden" is ignored by latest and greatest Safari 15.5. Wow. Nice! Good work!
-      // So use a fill/color of rgba(0,0,0,0)...
-      return svg`
-        <foreignObject width="0px" height="0px" x="${this.svg.xpx}" y="${this.svg.ypx}" overflow="hidden">
-          <body>
-            <div class="div__icon, hover" xmlns="http://www.w3.org/1999/xhtml"
-                style="line-height:${this.svg.iconPixels}px;position:relative;border-style:solid;border-width:0px;border-color:${this.alternateColor};fill:${this.alternateColor};color:${this.alternateColor};">
-                <ha-icon icon=${icon} id="icon-${this.toolId}"
-                @animationstart=${(e) => this._handleAnimationEvent(e, this)}
-                @animationiteration=${(e) => this._handleAnimationEvent(e, this)}
-                style="animation: flash 0.15s 20;"></ha-icon>
-            </div>
-          </body>
-        </foreignObject>
-        `;
-    }
-  }
-
-  _handleAnimationEvent(argEvent, argThis) {
-    argEvent.stopPropagation();
-    argEvent.preventDefault();
-
-    argThis.iconSvg = this._card.shadowRoot.getElementById('icon-'.concat(this.toolId))?.shadowRoot?.querySelectorAll('*')[0]?.path;
-    if (argThis.iconSvg) {
-      argThis._card.requestUpdate();
-    } else {
-    }
-  }
-
-  firstUpdated(changedProperties) {
-
-  }
-
-  /** *****************************************************************************
-  * EntityIconTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  * NTS:
-  * Adding        <style> div { overflow: hidden;}</style>
-  * to the <g group, clips the icon against the ha-card, ie the div.
-  * however, on Safari, all icons are clipped, as if they don't fit the room given to be displayed.
-  * a bug in rendering the Icon?? Only first time icon is clipped, then displayed normally if a data update
-  * from hass is coming in.
-  */
-
-  render() {
-    return svg`
-      <g "" id="icongrp-${this.toolId}" class="${classMap(this.classes.tool)}" style="${styleMap(this.styles.tool)}"
-        @click=${(e) => this.handleTapEvent(e, this.config)} >
-
-        ${this._renderIcon()}
-      </g>
-    `;
-  }
-} // END of class
-
-// -- End of Can't split these yet, due to circular dependencies...
-
-// ++ Class ++++++++++
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/** ***************************************************************************
-  * Toolset class
-  *
-  * Summary.
-  *
-  */
-
-class Toolset {
-  constructor(argCard, argConfig) {
-    this.toolsetId = Math.random().toString(36).substr(2, 9);
-    this._card = argCard;
-    this.dev = { ...this._card.dev };
-    if (this.dev.performance) console.time(`--> ${this.toolsetId} PERFORMANCE Toolset::constructor`);
-
-    this.config = argConfig;
-    this.tools = [];
-
-    // Get SVG coordinates.
-    this.svg = {};
-    this.svg.cx = Utils.calculateSvgCoordinate(argConfig.position.cx, SVG_DEFAULT_DIMENSIONS_HALF);
-    this.svg.cy = Utils.calculateSvgCoordinate(argConfig.position.cy, SVG_DEFAULT_DIMENSIONS_HALF);
-
-    this.svg.x = (this.svg.cx) - (SVG_DEFAULT_DIMENSIONS_HALF);
-    this.svg.y = (this.svg.cy) - (SVG_DEFAULT_DIMENSIONS_HALF);
-
-    // Group scaling experiment. Calc translate values for SVG using the toolset scale value
-    this.transform = {};
-    this.transform.scale = {};
-    this.transform.scale.x = this.transform.scale.y = 1;
-    this.transform.rotate = {};
-    this.transform.rotate.x = this.transform.rotate.y = 0;
-    this.transform.skew = {};
-    this.transform.skew.x = this.transform.skew.y = 0;
-
-    if (this.config.position.scale) {
-      this.transform.scale.x = this.transform.scale.y = this.config.position.scale;
-    }
-    if (this.config.position.rotate) {
-      this.transform.rotate.x = this.transform.rotate.y = this.config.position.rotate;
-    }
-
-    this.transform.scale.x = this.config.position.scale_x || this.config.position.scale || 1;
-    this.transform.scale.y = this.config.position.scale_y || this.config.position.scale || 1;
-
-    this.transform.rotate.x = this.config.position.rotate_x || this.config.position.rotate || 0;
-    this.transform.rotate.y = this.config.position.rotate_y || this.config.position.rotate || 0;
-
-    if (this.dev.debug) console.log('Toolset::constructor config/svg', this.toolsetId, this.config, this.svg);
-
-    // Create the tools configured in the toolset list.
-    const toolsNew = {
-      area: EntityAreaTool,
-      circslider: CircularSliderTool,
-      badge: BadgeTool,
-      bar: SparklineBarChartTool,
-      circle: CircleTool,
-      ellipse: EllipseTool,
-      horseshoe: HorseshoeTool,
-      icon: EntityIconTool,
-      line: LineTool,
-      name: EntityNameTool,
-      rectangle: RectangleTool,
-      rectex: RectangleToolEx,
-      regpoly: RegPolyTool,
-      segarc: SegmentedArcTool,
-      state: EntityStateTool,
-      slider: RangeSliderTool,
-      switch: SwitchTool,
-      text: TextTool,
-      usersvg: UserSvgTool,
-    };
-
-    this.config.tools.map((toolConfig) => {
-      const argConfig = { ...toolConfig };
-
-      const argPos = {
-        cx: 0 / 100 * SVG_DEFAULT_DIMENSIONS,
-        cy: 0 / 100 * SVG_DEFAULT_DIMENSIONS,
-        scale: this.config.position.scale ? this.config.position.scale : 1,
-      };
-
-      if (this.dev.debug) console.log('Toolset::constructor toolConfig', this.toolsetId, argConfig, argPos);
-
-      if (!toolConfig.disabled) {
-        const newTool = new toolsNew[toolConfig.type](this, argConfig, argPos);
-        this._card.entityHistory.needed |= (toolConfig.type === 'bar');
-        this.tools.push({ type: toolConfig.type, index: toolConfig.id, tool: newTool });
-      }
-    });
-
-    if (this.dev.performance) console.timeEnd(`--> ${this.toolsetId} PERFORMANCE Toolset::constructor`);
-  }
-
-  /** *****************************************************************************
-  * Toolset::updateValues()
-  *
-  * Summary.
-  * Called from set hass to update values for tools
-  *
-  */
-
-  // #TODO:
-  // Update only the changed entity_index, not all indexes. Now ALL tools are updated...
-  updateValues() {
-    if (this.dev.performance) console.time(`--> ${this.toolsetId} PERFORMANCE Toolset::updateValues`);
-    if (this.tools) {
-      this.tools.map((item, index) => {
-        if (true || item.type === 'segarc') {
-          if ((item.tool.config.hasOwnProperty('entity_index'))) {
-            if (this.dev.debug) console.log('Toolset::updateValues', item, index);
-            // if (this.dev.debug) console.log('Toolset::updateValues', typeof item.tool._stateValue);
-
-            // #IDEA @2021.11.20
-            // What if for attribute and secondaryinfo the entity state itself is also passsed automatically
-            // In that case that state is always present and can be used in animations by default.
-            // No need to pass an extra entity_index.
-            // A tool using the light brightness can also use the state (on/off) in that case for styling.
-            //
-            // Test can be done on 'state', 'attr', or 'secinfo' for default entity_index.
-            //
-            // Should pass a record in here orso as value { state : xx, attr: yy }
-
-            item.tool.value = this._card.attributesStr[item.tool.config.entity_index]
-              ? this._card.attributesStr[item.tool.config.entity_index]
-              : this._card.secondaryInfoStr[item.tool.config.entity_index]
-                ? this._card.secondaryInfoStr[item.tool.config.entity_index]
-                : this._card.entitiesStr[item.tool.config.entity_index];
-          }
-
-          // Check for multiple entities specified, and pass them to the tool
-          if ((item.tool.config.hasOwnProperty('entity_indexes'))) {
-            // Update list of entities in single record and pass that to the tool
-            // The first entity is used as the state, additional entities can help with animations,
-            // (used for formatting classes/styles) or can be used in a derived entity
-
-            const valueList = [];
-            for (let index = 0; index < item.tool.config.entity_indexes.length; ++index) {
-              valueList[index] = this._card.attributesStr[item.tool.config.entity_indexes[index].entity_index]
-                ? this._card.attributesStr[item.tool.config.entity_indexes[index].entity_index]
-                : this._card.secondaryInfoStr[item.tool.config.entity_indexes[index].entity_index]
-                  ? this._card.secondaryInfoStr[item.tool.config.entity_indexes[index].entity_index]
-                  : this._card.entitiesStr[item.tool.config.entity_indexes[index].entity_index];
-            }
-
-            item.tool.values = valueList;
-          }
-        }
-      });
-    }
-    if (this.dev.performance) console.timeEnd(`--> ${this.toolsetId} PERFORMANCE Toolset::updateValues`);
-  }
-
-  /** *****************************************************************************
-  * Toolset::connectedCallback()
-  *
-  * Summary.
-  *
-  */
-  connectedCallback() {
-    if (this.dev.performance) console.time(`--> ${this.toolsetId} PERFORMANCE Toolset::connectedCallback`);
-
-    if (this.dev.debug) console.log('*****Event - connectedCallback', this.toolsetId, new Date().getTime());
-    if (this.dev.performance) console.timeEnd(`--> ${this.toolsetId} PERFORMANCE Toolset::connectedCallback`);
-  }
-
-  /** *****************************************************************************
-  * Toolset::disconnectedCallback()
-  *
-  * Summary.
-  *
-  */
-  disconnectedCallback() {
-    if (this.dev.performance) console.time(`--> ${this.cardId} PERFORMANCE Toolset::disconnectedCallback`);
-
-    if (this.dev.debug) console.log('*****Event - disconnectedCallback', this.toolsetId, new Date().getTime());
-    if (this.dev.performance) console.timeEnd(`--> ${this.cardId} PERFORMANCE Toolset::disconnectedCallback`);
-  }
-
-  /** *****************************************************************************
-  * Toolset::firstUpdated()
-  *
-  * Summary.
-  *
-  */
-  firstUpdated(changedProperties) {
-    if (this.dev.debug) console.log('*****Event - Toolset::firstUpdated', this.toolsetId, new Date().getTime());
-
-    if (this.tools) {
-      this.tools.map((item, index) => {
-        if (typeof item.tool.firstUpdated === 'function') {
-          item.tool.firstUpdated(changedProperties);
-        }
-      });
-    }
-  }
-
-  /** *****************************************************************************
-  * Toolset::updated()
-  *
-  * Summary.
-  *
-  */
-  updated(changedProperties) {
-    if (this.dev.debug) console.log('*****Event - Updated', this.toolsetId, new Date().getTime());
-
-    if (this.tools) {
-      this.tools.map((item, index) => {
-        if (typeof item.tool.updated === 'function') {
-          item.tool.updated(changedProperties);
-        }
-      });
-    }
-  }
-
-  /** *****************************************************************************
-  * Toolset::renderToolset()
-  *
-  * Summary.
-  *
-  */
-  renderToolset() {
-    if (this.dev.debug) console.log('*****Event - renderToolset', this.toolsetId, new Date().getTime());
-
-    const svgItems = this.tools.map((item) => svg`
-          ${item.tool.render()}
-      `);
-    return svg`${svgItems}`;
-  }
-
-  /** *****************************************************************************
-  * Toolset::render()
-  *
-  * Summary.
-  * The render() function for this toolset renders all the tools within this set.
-  *
-  * Important notes:
-  * - the toolset position is set on the svg. That one accepts x,y
-  * - scaling, rotating and skewing (and translating) is done on the parent group.
-  *
-  * The order of transformations are done from the child's perspective!!
-  * So, the child (tools) gets positioned FIRST, and then scaled/rotated.
-  *
-  * See comments for different render paths for Apple/Safari and any other browser...
-  *
-  */
-
-  render() {
-    // Note:
-    // Rotating a card can produce different results on several browsers.
-    // A 1:1 card / toolset gives the same results, but other aspect ratio's may give different results.
-
-    if (((this._card.isSafari) || (this._card.iOS)) && (!this._card.isSafari16)) {
-      //
-      // Render path for Safari if not Safari 16:
-      //
-      // Safari seems to ignore - although not always - the transform-box:fill-box setting.
-      // - It needs the explicit center point when rotating. So this is added to the rotate() command.
-      // - scale around center uses the "move object to 0,0 -> scale -> move object back to position" trick,
-      //   where the second move takes scaling into account!
-      // - Does not apply transforms from the child's point of view.
-      //   Transform of toolset_position MUST take scaling of one level higher into account!
-      //
-      // Note: rotate is done around the defined center (cx,cy) of the toolsets position!
-      //
-      // More:
-      // - Safari NEEDS the overflow:visible on the <svg> element, as it defaults to "svg:{overflow: hidden;}".
-      //   Other browsers don't need that, they default to: "svg:not(:root) {overflow: hidden;}"
-      //
-      //   Without this setting, objects are cut-off or become invisible while scaled!
-
-      return svg`
-        <g id="toolset-${this.toolsetId}" class="toolset__group-outer"
-           transform="rotate(${this.transform.rotate.x}, ${this.svg.cx}, ${this.svg.cy})
-                      scale(${this.transform.scale.x}, ${this.transform.scale.y})
-                      "
-           style="transform-origin:center; transform-box:fill-box;">
-          <svg style="overflow:visible;">
-            <g class="toolset__group" transform="translate(${this.svg.cx / this.transform.scale.x}, ${this.svg.cy / this.transform.scale.y})">
-              ${this.renderToolset()}
-            </g>
-            </svg>
-        </g>
-      `;
-    } else {
-      //
-      // Render path for ANY other browser that usually follows the standards:
-      //
-      // - use transform-box:fill-box to make sure every transform is about the object itself!
-      // - applying the rules seen from the child's point of view.
-      //   So the transform on the toolset_position is NOT scaled, as the scaling is done one level higher.
-      //
-      // Note: rotate is done around the center of the bounding box. This might NOT be the toolsets center (cx,cy) position!
-      //
-      return svg`
-        <g id="toolset-${this.toolsetId}" class="toolset__group-outer"
-           transform="rotate(${this.transform.rotate.x}) scale(${this.transform.scale.x}, ${this.transform.scale.y})"
-           style="transform-origin:center; transform-box:fill-box;">
-          <svg style="overflow:visible;">
-            <g class="toolset__group" transform="translate(${this.svg.cx}, ${this.svg.cy})">
-              ${this.renderToolset()}
-            </g>
-            </svg>
-        </g>
-      `;
-    }
-  }
-} // END of class
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/**
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-//+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 // https://github.com/d3/d3-selection/blob/master/src/selection/data.js
 //
@@ -871,9 +166,6 @@ class SwissArmyKnifeCard extends LitElement {
       throw Error("card::constructor - Can't get Lovelace panel");
     }
 
-    if (!SwissArmyKnifeCard.sakIconCache) {
-      SwissArmyKnifeCard.sakIconCache = {};
-    }
     if (!SwissArmyKnifeCard.colorCache) {
       SwissArmyKnifeCard.colorCache = [];
     }
@@ -881,27 +173,6 @@ class SwissArmyKnifeCard extends LitElement {
     if (this.dev.debug) console.log('*****Event - card - constructor', this.cardId, new Date().getTime());
   }
 
-  /** *****************************************************************************
-  * Summary.
-  * Implements the properties method
-  *
-  */
-  /*
-  static get properties() {
-    return {
-      hass: {},
-      config: {},
-      states: [],
-      statesStr: [],
-
-      dashArray: String,
-      color1_offset: String,
-      color0: String,
-      color1: String,
-      angleCoords: Object
-    }
-  }
-*/
   static getSystemStyles() {
     return css`
       :host {
